build script for csharp build.
notes, this is for apache ant 1.7 or 1.8, which moves out .net related ant task to Apache Ant - Apache .NET Ant Library

this build should be only used in Windows box, with the corresponding .NET Framework SDK installed.

currently, in prod, v2.0 is used. and the environment 32bit, so the build should be on the same machine.